Make css separated and much more cleaner
