+++ backend/data/assignments.json
@@ -1,15 +1,15 @@
 [
   {
-    "id": "aeuu4aw66mcc56r4i",
-    "title": "adad",
-    "description": "adadadaadadad",
-    "subject": "adadadad",
-    "dueDate": "2025-11-11T07:11:00.000Z",
+    "id": "ti1jjukf2mccavxd3",
+    "title": "adadada",
+    "description": "adadada",
+    "subject": "dadadadad",
+    "dueDate": "2025-11-11T07:01:00.000Z",
     "priority": "medium",
-    "status": "completed",
+    "status": "not-started",
     "estimatedHours": 1,
     "tags": [],
-    "createdAt": "2025-06-25T16:01:45.426Z",
-    "updatedAt": "2025-06-25T16:01:50.802Z"
+    "createdAt": "2025-06-25T18:41:17.991Z",
+    "updatedAt": "2025-06-25T18:41:17.991Z"
   }
 ]
