+++ frontend/src/styles/components/buttons.css
@@ -1,0 +1,101 @@
+/* Buttons */
+.btn {
+	display: inline-block;
+	padding: 0.5rem 1rem;
+	border-radius: 0.375rem;
+	border: none;
+	font-weight: 500;
+	text-decoration: none;
+	cursor: pointer;
+	transition: all 0.2s;
+}
+
+.btn:disabled {
+	opacity: 0.6;
+	cursor: not-allowed;
+}
+
+.btn-primary {
+	background-color: #2563eb;
+	color: white;
+}
+
+.btn-primary:hover {
+	background-color: #1d4ed8;
+}
+
+.btn-secondary {
+	background-color: #6b7280;
+	color: white;
+}
+
+.btn-secondary:hover {
+	background-color: #4b5563;
+}
+
+.btn-success {
+	background-color: #059669;
+	color: white;
+}
+
+.btn-success:hover {
+	background-color: #047857;
+}
+
+.btn-danger {
+	background-color: #dc2626;
+	color: white;
+}
+
+.btn-danger:hover {
+	background-color: #b91c1c;
+}
+
+.btn-sm {
+	padding: 0.375rem 0.75rem;
+	font-size: 0.875rem;
+	line-height: 1.25rem;
+}
+
+/* Action Buttons */
+.action-btn {
+	padding: 0.25rem 0.5rem;
+	border: 1px solid #e5e7eb;
+	border-radius: 0.375rem;
+	background: white;
+	cursor: pointer;
+	transition: all 0.2s;
+}
+
+.action-btn:hover {
+	background-color: #f9fafb;
+}
+
+.edit-btn:hover {
+	border-color: #3b82f6;
+	background-color: #eff6ff;
+}
+
+.delete-btn {
+	transition: all 0.2s;
+}
+
+.delete-btn:hover {
+	background-color: #fef2f2;
+	border-color: #f87171;
+	transform: scale(1.1);
+}
+
+/* Sort Direction Button */
+.sort-direction-btn {
+	padding: 0.25rem 0.5rem;
+	font-size: 0.875rem;
+	border: 1px solid #d1d5db;
+	border-radius: 0.375rem;
+	background: white;
+	cursor: pointer;
+}
+
+.sort-direction-btn:hover {
+	background-color: #f9fafb;
+}
