+++ frontend/src/styles/components/cards.css
@@ -1,0 +1,121 @@
+/* Cards */
+.card {
+	background: white;
+	border-radius: 0.5rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+	padding: 1.5rem;
+	margin-bottom: 1.5rem;
+}
+
+.card:hover {
+	box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+	transition: box-shadow 0.2s;
+}
+
+.card-grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+	gap: 1.5rem;
+}
+
+/* Empty State */
+.empty-state {
+	text-align: center;
+	padding: 3rem 1.5rem;
+}
+
+.empty-icon {
+	font-size: 4rem;
+	color: #d1d5db;
+	margin-bottom: 1rem;
+}
+
+.empty-title {
+	font-size: 1.25rem;
+	font-weight: 600;
+	color: #111827;
+	margin-bottom: 0.5rem;
+}
+
+.empty-description {
+	color: #6b7280;
+}
+
+/* Priority Badges */
+.priority-badge {
+	padding: 0.25rem 0.75rem;
+	border-radius: 9999px;
+	font-size: 0.75rem;
+	font-weight: 600;
+	white-space: nowrap;
+}
+
+.priority-high {
+	background-color: #fee2e2;
+	color: #991b1b;
+}
+
+.priority-medium {
+	background-color: #fef3c7;
+	color: #92400e;
+}
+
+.priority-low {
+	background-color: #dcfce7;
+	color: #166534;
+}
+
+.priority-default {
+	background-color: #f3f4f6;
+	color: #1f2937;
+}
+
+/* Status Badges */
+.status-select {
+	padding: 0.25rem 0.5rem;
+	border-radius: 0.375rem;
+	font-size: 0.75rem;
+	font-weight: 500;
+	cursor: pointer;
+}
+
+.status-completed {
+	background-color: #dcfce7;
+	color: #166534;
+	border: 1px solid #bbf7d0;
+}
+
+.status-in-progress {
+	background-color: #dbeafe;
+	color: #1e40af;
+	border: 1px solid #bfdbfe;
+}
+
+.status-not-started {
+	background-color: #f3f4f6;
+	color: #1f2937;
+	border: 1px solid #e5e7eb;
+}
+
+.status-overdue {
+	background-color: #fee2e2;
+	color: #991b1b;
+	border: 1px solid #fecaca;
+}
+
+/* Due Date Styles */
+.due-date-completed {
+	color: #059669;
+}
+
+.due-date-overdue {
+	color: #dc2626;
+}
+
+.due-date-urgent {
+	color: #ea580c;
+}
+
+.due-date-normal {
+	color: #6b7280;
+}
