+++ frontend/src/styles/components/errors.css
@@ -1,0 +1,121 @@
+/* Error Components */
+
+/* Error Boundary */
+.error-boundary {
+	padding: 2rem;
+	text-align: center;
+	min-height: 400px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+.error-boundary > div {
+	animation: errorSlideIn 0.3s ease-out;
+}
+
+/* Error Fallback */
+.error-fallback {
+	min-height: 300px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
+
+/* Network Error */
+.network-error {
+	margin-bottom: 1rem;
+	border-left: 4px solid #ef4444;
+	animation: errorSlideIn 0.3s ease-out;
+}
+
+.network-error .text-red-800 {
+	color: #991b1b;
+}
+
+.network-error .text-red-700 {
+	color: #b91c1c;
+}
+
+.network-error .text-red-600 {
+	color: #dc2626;
+}
+
+.network-error .bg-red-50 {
+	background-color: #fef2f2;
+}
+
+.network-error .border-red-200 {
+	border-color: #fecaca;
+}
+
+.network-error .bg-red-100 {
+	background-color: #fee2e2;
+}
+
+.network-error .bg-red-200 {
+	background-color: #fecaca;
+}
+
+/* Error Details */
+details[open] summary {
+	margin-bottom: 0.5rem;
+}
+
+details summary {
+	list-style: none;
+	padding: 0.5rem;
+	border-radius: 0.375rem;
+	background-color: rgba(0, 0, 0, 0.05);
+}
+
+details summary::-webkit-details-marker {
+	display: none;
+}
+
+details summary::before {
+	content: "▶ ";
+	display: inline-block;
+	transition: transform 0.2s ease;
+}
+
+details[open] summary::before {
+	transform: rotate(90deg);
+}
+
+/* Pre-formatted Text in Error Details */
+.network-error pre,
+.error-boundary pre {
+	font-family: "Courier New", monospace;
+	font-size: 0.75rem;
+	line-height: 1.4;
+	max-height: 200px;
+	overflow-y: auto;
+}
+
+/* Error Animations */
+@keyframes errorSlideIn {
+	from {
+		opacity: 0;
+		transform: translateY(-10px);
+	}
+	to {
+		opacity: 1;
+		transform: translateY(0);
+	}
+}
+
+/* Error Button Hover Effects */
+.network-error button:hover,
+.error-boundary button:hover {
+	transform: translateY(-1px);
+	transition: transform 0.2s ease;
+}
+
+/* Focus States for Error Buttons */
+.error-boundary button:focus,
+.network-error button:focus,
+.error-fallback button:focus {
+	outline: 2px solid #3b82f6;
+	outline-offset: 2px;
+}
