+++ frontend/src/styles/features/assignments.css
@@ -1,0 +1,140 @@
+/* Assignment-specific styles */
+
+/* Assignment List */
+.assignment-list {
+	max-width: 100%;
+}
+
+.assignment-controls {
+	display: flex;
+	flex-wrap: wrap;
+	gap: 1rem;
+	align-items: center;
+	margin-bottom: 1.5rem;
+	padding: 1rem;
+	background: white;
+	border-radius: 0.5rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+}
+
+.control-group {
+	display: flex;
+	align-items: center;
+	gap: 0.5rem;
+}
+
+.control-label {
+	font-size: 0.875rem;
+	font-weight: 500;
+	color: #374151;
+}
+
+.results-count {
+	font-size: 0.875rem;
+	color: #6b7280;
+	margin-left: auto;
+}
+
+/* Assignment Cards */
+.assignment-cards {
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+}
+
+.assignment-card {
+	background: white;
+	border-radius: 0.5rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+	padding: 1.5rem;
+	transition: box-shadow 0.2s ease;
+}
+
+.assignment-card:hover {
+	box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+/* Assignment Header */
+.assignment-header {
+	display: flex;
+	align-items: flex-start;
+	justify-content: space-between;
+	margin-bottom: 1rem;
+}
+
+.assignment-title-section {
+	flex: 1;
+	min-width: 0;
+}
+
+.assignment-title {
+	font-size: 1.125rem;
+	font-weight: 600;
+	color: #111827;
+	margin: 0 0 0.25rem 0;
+}
+
+.assignment-subject {
+	font-size: 0.875rem;
+	color: #6b7280;
+	margin: 0;
+}
+
+/* Assignment Description */
+.assignment-description {
+	color: #374151;
+	margin-bottom: 1rem;
+	line-height: 1.5;
+}
+
+/* Assignment Tags */
+.assignment-tags {
+	display: flex;
+	flex-wrap: wrap;
+	gap: 0.25rem;
+	margin-bottom: 1rem;
+}
+
+/* Assignment Footer */
+.assignment-footer {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	padding-top: 1rem;
+	border-top: 1px solid #f3f4f6;
+}
+
+.assignment-meta {
+	display: flex;
+	flex-direction: column;
+	gap: 0.25rem;
+	font-size: 0.875rem;
+}
+
+.due-date {
+	font-weight: 500;
+}
+
+.time-info {
+	color: #6b7280;
+}
+
+.assignment-actions {
+	display: flex;
+	align-items: center;
+	gap: 0.75rem;
+}
+
+/* Action Buttons */
+.action-buttons {
+	display: flex;
+	gap: 0.25rem;
+}
+
+/* Assignment Form */
+.assignment-form h2 {
+	color: #2563eb;
+	border-bottom: 2px solid #e5e7eb;
+	padding-bottom: 0.5rem;
+	margin-bottom: 1.5rem;
+}
