+++ frontend/src/styles/features/dashboard.css
@@ -1,0 +1,415 @@
+/* Dashboard-specific styles */
+
+/* Dashboard Layout */
+.dashboard {
+	max-width: 1200px;
+	margin: 0 auto;
+	padding: 0 1rem;
+}
+
+.dashboard-header {
+	margin-bottom: 2rem;
+	text-align: center;
+}
+
+.dashboard-header .flex {
+	justify-content: center;
+	flex-wrap: wrap;
+	gap: 1rem;
+}
+
+.dashboard-title {
+	font-size: 2.5rem;
+	font-weight: bold;
+	color: #111827;
+	margin-bottom: 0.5rem;
+}
+
+.dashboard-subtitle {
+	font-size: 1.125rem;
+	color: #6b7280;
+	margin: 0;
+}
+
+/* Stats Section */
+.stats-section {
+	margin-bottom: 3rem;
+}
+
+.section-title {
+	font-size: 1.5rem;
+	font-weight: 600;
+	color: #111827;
+	margin-bottom: 1.5rem;
+}
+
+.stats-grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+	gap: 1.5rem;
+	margin-bottom: 2rem;
+}
+
+.stat-card {
+	background: white;
+	padding: 1.5rem;
+	border-radius: 0.75rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+	border: 1px solid #f3f4f6;
+	display: flex;
+	align-items: center;
+	gap: 1rem;
+	transition: all 0.2s;
+}
+
+.stat-card:hover {
+	transform: translateY(-2px);
+	box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+	border-color: #e5e7eb;
+}
+
+.stat-icon {
+	font-size: 2rem;
+	line-height: 1;
+}
+
+.stat-content {
+	flex: 1;
+}
+
+.stat-number {
+	font-size: 2rem;
+	font-weight: bold;
+	color: #111827;
+	line-height: 1.2;
+}
+
+.stat-label {
+	font-size: 0.875rem;
+	color: #6b7280;
+	font-weight: 500;
+	margin-top: 0.25rem;
+}
+
+/* Progress Section */
+.progress-section {
+	background: white;
+	padding: 1.5rem;
+	border-radius: 0.75rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+	border: 1px solid #f3f4f6;
+}
+
+.progress-header {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	margin-bottom: 1rem;
+}
+
+.progress-title {
+	font-size: 1.125rem;
+	font-weight: 600;
+	color: #111827;
+	margin: 0;
+}
+
+.progress-percentage {
+	font-size: 1.5rem;
+	font-weight: bold;
+	color: #2563eb;
+}
+
+.progress-bar {
+	width: 100%;
+	height: 1rem;
+	background-color: #f3f4f6;
+	border-radius: 0.5rem;
+	overflow: hidden;
+	margin-bottom: 0.75rem;
+}
+
+.progress-fill {
+	height: 100%;
+	background: linear-gradient(90deg, #3b82f6, #1d4ed8);
+	border-radius: 0.5rem;
+	transition: width 0.3s ease;
+}
+
+.progress-details {
+	text-align: center;
+}
+
+.progress-detail {
+	font-size: 0.875rem;
+	color: #6b7280;
+}
+
+/* Dashboard Grid */
+.dashboard-grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
+	gap: 1.5rem;
+}
+
+/* Dashboard Cards */
+.dashboard-card {
+	background: white;
+	border-radius: 0.75rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+	border: 1px solid #f3f4f6;
+	overflow: hidden;
+	transition: all 0.2s;
+}
+
+.dashboard-card:hover {
+	box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+	border-color: #e5e7eb;
+}
+
+.dashboard-card.urgent {
+	border-color: #fca5a5;
+	background: linear-gradient(135deg, #fef2f2 0%, white 100%);
+}
+
+.card-header {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	padding: 1.5rem 1.5rem 1rem 1.5rem;
+	border-bottom: 1px solid #f3f4f6;
+}
+
+.card-title {
+	font-size: 1.125rem;
+	font-weight: 600;
+	color: #111827;
+	margin: 0;
+}
+
+.card-link {
+	font-size: 0.875rem;
+	color: #2563eb;
+	text-decoration: none;
+	font-weight: 500;
+	transition: color 0.2s;
+}
+
+.card-link:hover {
+	color: #1d4ed8;
+}
+
+/* Assignment Items */
+.assignments-list {
+	padding: 1rem 1.5rem 1.5rem 1.5rem;
+}
+
+.assignment-item {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	padding: 0.75rem 0;
+	border-bottom: 1px solid #f9fafb;
+	transition: all 0.2s;
+}
+
+.assignment-item:last-child {
+	border-bottom: none;
+}
+
+.assignment-item:hover {
+	background-color: #f9fafb;
+	margin: 0 -0.75rem;
+	padding-left: 0.75rem;
+	padding-right: 0.75rem;
+	border-radius: 0.5rem;
+}
+
+.assignment-item.overdue {
+	background-color: #fef2f2;
+	border-color: #fecaca;
+	border-radius: 0.5rem;
+	margin: 0 -0.75rem 0.5rem -0.75rem;
+	padding-left: 0.75rem;
+	padding-right: 0.75rem;
+}
+
+.assignment-content {
+	flex: 1;
+	min-width: 0;
+}
+
+.assignment-title {
+	font-weight: 600;
+	color: #111827;
+	font-size: 0.875rem;
+	margin-bottom: 0.25rem;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+}
+
+.assignment-subject {
+	font-size: 0.75rem;
+	color: #6b7280;
+}
+
+.assignment-meta {
+	display: flex;
+	align-items: center;
+	gap: 0.5rem;
+}
+
+.assignment-priority {
+	font-size: 1rem;
+}
+
+.assignment-due {
+	font-size: 0.75rem;
+	color: #6b7280;
+	font-weight: 500;
+}
+
+.overdue-text {
+	color: #dc2626 !important;
+	font-weight: 600;
+}
+
+/* Due Date Warnings */
+.due-warning {
+	font-size: 0.625rem;
+	font-weight: 700;
+	padding: 0.125rem 0.375rem;
+	border-radius: 0.25rem;
+	margin-top: 0.25rem;
+	display: inline-block;
+}
+
+.warning-overdue {
+	background-color: #fef2f2;
+	color: #dc2626;
+	border: 1px solid #fecaca;
+}
+
+.warning-today {
+	background-color: #fef3c7;
+	color: #d97706;
+	border: 1px solid #fde68a;
+}
+
+.warning-tomorrow {
+	background-color: #fef3c7;
+	color: #d97706;
+	border: 1px solid #fde68a;
+}
+
+.warning-soon {
+	background-color: #eff6ff;
+	color: #2563eb;
+	border: 1px solid #bfdbfe;
+}
+
+/* Notes List */
+.notes-list {
+	padding: 1rem 1.5rem 1.5rem 1.5rem;
+}
+
+.note-item {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	padding: 0.75rem 0;
+	border-bottom: 1px solid #f9fafb;
+	transition: all 0.2s;
+}
+
+.note-item:last-child {
+	border-bottom: none;
+}
+
+.note-item:hover {
+	background-color: #f9fafb;
+	margin: 0 -0.75rem;
+	padding-left: 0.75rem;
+	padding-right: 0.75rem;
+	border-radius: 0.5rem;
+}
+
+.note-content {
+	flex: 1;
+	min-width: 0;
+}
+
+.note-title {
+	font-weight: 600;
+	color: #111827;
+	font-size: 0.875rem;
+	margin-bottom: 0.25rem;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+}
+
+.note-subject {
+	font-size: 0.75rem;
+	color: #6b7280;
+}
+
+.note-date {
+	font-size: 0.75rem;
+	color: #6b7280;
+	font-weight: 500;
+}
+
+/* Empty Message */
+.empty-message {
+	padding: 2rem 1.5rem;
+	text-align: center;
+}
+
+.empty-icon {
+	font-size: 3rem;
+	margin-bottom: 0.5rem;
+}
+
+.empty-message p {
+	color: #6b7280;
+	margin-bottom: 1rem;
+}
+
+/* Quick Actions */
+.quick-actions {
+	padding: 1rem 1.5rem 1.5rem 1.5rem;
+	display: grid;
+	gap: 0.75rem;
+}
+
+.quick-action {
+	display: flex;
+	align-items: center;
+	gap: 1rem;
+	padding: 0.75rem;
+	border-radius: 0.5rem;
+	background-color: #f9fafb;
+	text-decoration: none;
+	color: #374151;
+	font-weight: 500;
+	transition: all 0.2s;
+	border: none;
+	cursor: pointer;
+	font-size: 0.875rem;
+}
+
+.quick-action:hover {
+	background-color: #f3f4f6;
+	color: #111827;
+	transform: translateX(2px);
+}
+
+.action-icon {
+	font-size: 1.25rem;
+}
+
+.action-text {
+	flex: 1;
+}
