+++ frontend/src/styles/features/notes.css
@@ -1,0 +1,145 @@
+/* Notes-specific styles */
+
+/* Note Form */
+.note-form {
+	max-width: 800px;
+	margin: 0 auto;
+}
+
+/* Note List */
+.note-list {
+	max-width: 100%;
+}
+
+.note-controls {
+	display: flex;
+	flex-wrap: wrap;
+	gap: 1rem;
+	align-items: center;
+	margin-bottom: 1.5rem;
+	padding: 1rem;
+	background: white;
+	border-radius: 0.5rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+}
+
+/* Notes Grid */
+.notes-grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fill, minmax(350px, 1fr));
+	gap: 1.5rem;
+}
+
+.note-card {
+	background: white;
+	border-radius: 0.5rem;
+	box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+	padding: 1.5rem;
+	transition: all 0.2s ease;
+	border-left: 4px solid transparent;
+}
+
+.note-card:hover {
+	transform: translateY(-2px);
+	box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+	border-left-color: #059669;
+}
+
+/* Note Header */
+.note-header {
+	display: flex;
+	align-items: flex-start;
+	justify-content: space-between;
+	margin-bottom: 1rem;
+}
+
+.note-title-section {
+	flex: 1;
+	min-width: 0;
+}
+
+.note-title {
+	font-size: 1.125rem;
+	font-weight: 600;
+	color: #111827;
+	margin: 0 0 0.25rem 0;
+	line-height: 1.3;
+}
+
+.note-subject {
+	font-size: 0.875rem;
+	color: #6b7280;
+	margin: 0;
+}
+
+.note-actions {
+	display: flex;
+	gap: 0.25rem;
+}
+
+/* Linked Assignment */
+.linked-assignment {
+	display: flex;
+	align-items: center;
+	gap: 0.5rem;
+	margin-bottom: 1rem;
+	padding: 0.5rem;
+	background-color: #eff6ff;
+	border-radius: 0.375rem;
+	border: 1px solid #bfdbfe;
+}
+
+.link-icon {
+	font-size: 0.875rem;
+}
+
+.assignment-link {
+	font-size: 0.875rem;
+	color: #1d4ed8;
+	font-weight: 500;
+}
+
+/* Note Content */
+.note-content {
+	color: #374151;
+	line-height: 1.6;
+	margin-bottom: 1rem;
+	white-space: pre-wrap;
+	word-wrap: break-word;
+}
+
+/* Note Tags */
+.note-tags {
+	display: flex;
+	flex-wrap: wrap;
+	gap: 0.25rem;
+	margin-bottom: 1rem;
+}
+
+.note-tags .tag {
+	padding: 0.25rem 0.5rem;
+	font-size: 0.75rem;
+	background-color: #dcfce7;
+	color: #166534;
+	border-radius: 9999px;
+}
+
+/* Note Footer */
+.note-footer {
+	padding-top: 1rem;
+	border-top: 1px solid #f3f4f6;
+}
+
+.note-dates {
+	display: flex;
+	flex-direction: column;
+	gap: 0.25rem;
+}
+
+.note-date {
+	font-size: 0.75rem;
+	color: #6b7280;
+	display: flex;
+	align-items: center;
+	gap: 0.25rem;
+}
