+++ frontend/src/styles/globals/typography.css
@@ -1,0 +1,69 @@
+/* Typography */
+.text-center {
+	text-align: center;
+}
+
+.text-sm {
+	font-size: 0.875rem;
+}
+
+.text-lg {
+	font-size: 1.125rem;
+}
+
+.text-xl {
+	font-size: 1.25rem;
+}
+
+.text-2xl {
+	font-size: 1.5rem;
+}
+
+.text-3xl {
+	font-size: 1.875rem;
+}
+
+.text-4xl {
+	font-size: 2.25rem;
+}
+
+.font-bold {
+	font-weight: bold;
+}
+
+.font-semibold {
+	font-weight: 600;
+}
+
+/* Text Colors */
+.text-gray-600 {
+	color: #4b5563;
+}
+
+.text-gray-900 {
+	color: #111827;
+}
+
+.text-blue-600 {
+	color: #2563eb;
+}
+
+.text-green-600 {
+	color: #059669;
+}
+
+.text-yellow-600 {
+	color: #d97706;
+}
+
+.text-red-600 {
+	color: #dc2626;
+}
+
+.text-purple-600 {
+	color: #9333ea;
+}
+
+.text-orange-600 {
+	color: #ea580c;
+}
