+++ frontend/src/styles/globals/utilities.css
@@ -1,0 +1,159 @@
+/* Utility Classes */
+
+/* Spacing */
+.mb-2 {
+	margin-bottom: 0.5rem;
+}
+.mb-4 {
+	margin-bottom: 1rem;
+}
+.mb-6 {
+	margin-bottom: 1.5rem;
+}
+.mb-8 {
+	margin-bottom: 2rem;
+}
+.mt-2 {
+	margin-top: 0.5rem;
+}
+.mt-4 {
+	margin-top: 1rem;
+}
+.mt-6 {
+	margin-top: 1.5rem;
+}
+.ml-1 {
+	margin-left: 0.25rem;
+}
+.mt-1 {
+	margin-top: 0.25rem;
+}
+
+.px-2 {
+	padding-left: 0.5rem;
+	padding-right: 0.5rem;
+}
+.py-1 {
+	padding-top: 0.25rem;
+	padding-bottom: 0.25rem;
+}
+
+.gap-2 {
+	gap: 0.5rem;
+}
+.gap-4 {
+	gap: 1rem;
+}
+
+/* Flex utilities */
+.flex {
+	display: flex;
+}
+.flex-between {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+}
+.flex-center {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+.flex-wrap {
+	flex-wrap: wrap;
+}
+.flex-1 {
+	flex: 1;
+}
+.inline-flex {
+	display: inline-flex;
+}
+.items-center {
+	align-items: center;
+}
+
+/* Background Colors */
+.bg-blue-100 {
+	background-color: #dbeafe;
+}
+.bg-red-500 {
+	background-color: #ef4444;
+}
+.bg-yellow-500 {
+	background-color: #eab308;
+}
+.bg-green-500 {
+	background-color: #22c55e;
+}
+.bg-yellow-50 {
+	background-color: #fefce8;
+}
+.bg-orange-50 {
+	background-color: #fff7ed;
+}
+
+/* Border Colors */
+.border-red-500 {
+	border-color: #ef4444 !important;
+}
+.border-yellow-200 {
+	border-color: #fde047;
+}
+.border-orange-200 {
+	border-color: #fed7aa;
+}
+
+/* Text Colors (additional) */
+.text-blue-800 {
+	color: #1e40af;
+}
+.text-blue-600 {
+	color: #2563eb;
+}
+.text-yellow-600 {
+	color: #ca8a04;
+}
+.text-yellow-800 {
+	color: #854d0e;
+}
+.text-orange-600 {
+	color: #ea580c;
+}
+.text-orange-800 {
+	color: #9a3412;
+}
+
+/* Hover Effects */
+.hover\:text-blue-800:hover {
+	color: #1e40af;
+}
+
+/* Border radius */
+.rounded {
+	border-radius: 0.25rem;
+}
+
+/* Z-index */
+.z-50 {
+	z-index: 50;
+}
+
+/* Width utilities */
+.min-w-24 {
+	min-width: 6rem;
+}
+
+/* Transition utilities */
+.transition-colors {
+	transition-property: color, background-color, border-color;
+	transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+	transition-duration: 150ms;
+}
+
+/* Animation utilities */
+.animate-spin {
+	animation: spin 1s linear infinite;
+}
+.animate-pulse {
+	animation: pulse 2s cubic-bezier(0.4, 0, 0.6, 1) infinite;
+}
