+++ frontend/src/styles/globals/variables.css
@@ -1,0 +1,56 @@
+/* CSS Custom Properties */
+:root {
+	/* Colors */
+	--color-primary: #2563eb;
+	--color-primary-dark: #1d4ed8;
+	--color-primary-light: #3b82f6;
+
+	--color-secondary: #6b7280;
+	--color-secondary-dark: #4b5563;
+
+	--color-success: #059669;
+	--color-success-dark: #047857;
+
+	--color-danger: #dc2626;
+	--color-danger-dark: #b91c1c;
+
+	--color-warning: #d97706;
+	--color-warning-dark: #92400e;
+
+	/* Text Colors */
+	--color-text-primary: #111827;
+	--color-text-secondary: #6b7280;
+	--color-text-muted: #9ca3af;
+
+	/* Background Colors */
+	--color-bg-primary: #ffffff;
+	--color-bg-secondary: #f9fafb;
+	--color-bg-muted: #f3f4f6;
+
+	/* Border Colors */
+	--color-border: #e5e7eb;
+	--color-border-light: #f3f4f6;
+	--color-border-dark: #d1d5db;
+
+	/* Spacing */
+	--spacing-xs: 0.25rem;
+	--spacing-sm: 0.5rem;
+	--spacing-md: 1rem;
+	--spacing-lg: 1.5rem;
+	--spacing-xl: 2rem;
+
+	/* Border Radius */
+	--radius-sm: 0.25rem;
+	--radius-md: 0.375rem;
+	--radius-lg: 0.5rem;
+	--radius-xl: 0.75rem;
+
+	/* Shadows */
+	--shadow-sm: 0 1px 3px rgba(0, 0, 0, 0.1);
+	--shadow-md: 0 4px 6px rgba(0, 0, 0, 0.1);
+	--shadow-lg: 0 10px 15px rgba(0, 0, 0, 0.1);
+
+	/* Transitions */
+	--transition-fast: 0.2s ease;
+	--transition-normal: 0.3s ease;
+}
