+++ frontend/src/styles/layout/navigation.css
@@ -1,0 +1,48 @@
+/* Navigation */
+.nav {
+	background-color: #2563eb;
+	color: white;
+	box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.nav-container {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	height: 4rem;
+	max-width: 1200px;
+	margin: 0 auto;
+	padding: 0 1rem;
+}
+
+.nav-brand {
+	font-size: 1.25rem;
+	font-weight: bold;
+	text-decoration: none;
+	color: white;
+}
+
+.nav-links {
+	display: flex;
+	gap: 1rem;
+	list-style: none;
+}
+
+.nav-link {
+	padding: 0.5rem 0.75rem;
+	border-radius: 0.375rem;
+	text-decoration: none;
+	color: #dbeafe;
+	font-weight: 500;
+	transition: all 0.2s;
+}
+
+.nav-link:hover {
+	background-color: #3b82f6;
+	color: white;
+}
+
+.nav-link.active {
+	background-color: #1d4ed8;
+	color: white;
+}
